Move MailHelpers from /lib to /app/mailers/concerns
