+++ app/mailers/visitor_mailer.rb
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 class VisitorMailer < ActionMailer::Base
-  include MailerHelper::NoReply
-  include MailerHelper::Autoresponder
-  include MailerHelper::Addresses
+  include NoReply
+  include Autoresponder
+  include Addresses
   include EnsureQuotedPrintable
 
   add_template_helper(ApplicationHelper)
